[DIDA] fix. NFT 만들기 > NFT 만들기 로직 수정
